CuoiKy/final_project/requirements.txt
absl-py
astunparse
black
cachetools
certifi
charset-normalizer
click
colorama
cycler
flatbuffers
fonttools
gast
google-auth
google-auth-oauthlib
google-pasta
grpcio
h5py
idna
joblib
keras
keras-facenet
keras-nightly
Keras-Preprocessing
kiwisolver
libclang
Markdown
matplotlib
mtcnn
mypy-extensions
numpy
oauthlib
opencv-python
opt-einsum
packaging
pathspec
Pillow
platformdirs
protobuf
psutil
pyasn1
pyasn1-modules
pyparsing
python-dateutil
requests
requests-oauthlib
rsa
scikit-learn
scipy
six
tb-nightly
tensorboard
tensorboard-data-server
tensorboard-plugin-wit
tensorflow
tensorflow-estimator
tensorflow-io-gcs-filesystem
termcolor
tf-estimator-nightly
tf-nightly-gpu
threadpoolctl
tomli
typing_extensions 
urllib3
Werkzeug
wrapt
